refactor: restructure UI navigation
