+++ services/ui/app/explore/layout.tsx
@@ -1,0 +1,10 @@
+import ExploreTabs from '../../components/ExploreTabs';
+
+export default function ExploreLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <section className="space-y-6">
+      <ExploreTabs />
+      {children}
+    </section>
+  );
+}
