+++ services/ui/app/explore/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
 
-export default function Page() {
+export default function ExploreIndex() {
   redirect('/explore/trajectory');
 }
