+++ services/ui/app/explore/radar/loading.tsx
@@ -1,6 +1,6 @@
-import ChartSkeleton from '../../components/ChartSkeleton';
-import ChartContainer from '../../components/ChartContainer';
-import Skeleton from '../../components/Skeleton';
+import ChartSkeleton from '../../../components/ChartSkeleton';
+import ChartContainer from '../../../components/ChartContainer';
+import Skeleton from '../../../components/Skeleton';
 
 export default function Loading() {
   return (
