+++ services/ui/app/insights/error.tsx
