+++ services/ui/app/insights/loading.tsx
@@ -1,4 +1,4 @@
-import Skeleton from '../../../components/Skeleton';
+import Skeleton from '../../components/Skeleton';
 
 export default function Loading() {
   return (
